Moved browser cleanup code to browser util module; added cleanup step to execute (#192)
